[Mounts] Split mount related functions into their own file.
