+++ crawl-ref/source/MSVC/crawl.vcxproj.filters
@@ -58,6 +58,9 @@
     <ClCompile Include="..\butcher.cc">
       <Filter>cc</Filter>
     </ClCompile>
+	<ClCompile Include="..\chaos.cc">
+		<Filter>cc</Filter>
+	</ClCompile>
     <ClCompile Include="..\chardump.cc">
       <Filter>cc</Filter>
     </ClCompile>
@@ -913,6 +916,9 @@
     <ClCompile Include="..\movement.cc">
       <Filter>cc</Filter>
     </ClCompile>
+    <ClCompile Include="..\mount.cc">
+      <Filter>cc</Filter>
+    </ClCompile>
     <ClCompile Include="..\rltiles\tiledef-gui.cc">
       <Filter>rtiles_generated</Filter>
     </ClCompile>
@@ -1042,6 +1048,9 @@
       <Filter>h</Filter>
     </ClInclude>
     <ClInclude Include="..\canned-message-type.h">
+      <Filter>h</Filter>
+    </ClInclude>
+    <ClInclude Include="..\chaos.h">
       <Filter>h</Filter>
     </ClInclude>
     <ClInclude Include="..\chardump.h">
@@ -2245,6 +2254,9 @@
       <Filter>h</Filter>
     </ClInclude>
     <ClInclude Include="..\wu-jian-attack-type.h">
+      <Filter>h</Filter>
+    </ClInclude>
+    <ClInclude Include="..\mount.h">
       <Filter>h</Filter>
     </ClInclude>
     <ClInclude Include="..\movement.h">
