+++ crawl-ref/source/Makefile.obj
@@ -164,6 +164,7 @@ mon-tentacle.o \
 mon-transit.o \
 mon-util.o \
 monster.o \
+mount.o \
 movement.o \
 mutation.o \
 nearby-danger.o \
