+++ crawl-ref/source/android-project/jni/src/Android.mk
@@ -36,6 +36,7 @@ LOCAL_SRC_FILES := $(SDL_PATH)/src/main/android/SDL_android_main.c \
     $(CRAWL_PATH)/branch.cc \
     $(CRAWL_PATH)/butcher.cc \
     $(CRAWL_PATH)/bloodspatter.cc \
+	$(CRAWL_PATH)/chaos.cc \
     $(CRAWL_PATH)/chardump.cc \
     $(CRAWL_PATH)/cio.cc \
     $(CRAWL_PATH)/cloud.cc \
@@ -182,6 +183,7 @@ LOCAL_SRC_FILES := $(SDL_PATH)/src/main/android/SDL_android_main.c \
     $(CRAWL_PATH)/mon-transit.cc \
     $(CRAWL_PATH)/mon-util.cc \
     $(CRAWL_PATH)/monster.cc \
+	$(CRAWL_PATH)/mount.cc \
     $(CRAWL_PATH)/mutation.cc \
     $(CRAWL_PATH)/nearby-danger.cc \
     $(CRAWL_PATH)/newgame.cc \
