+++ crawl-ref/source/spl-other.cc
@@ -529,118 +529,4 @@ spret cast_SMD(const coord_def& target, int pow, bool fail)
     return spret::success;
 }
 
-spret gain_mount(mount_type mount, int pow, bool fail)
-{
-    if (you.transform_uncancellable)
-    {
-        mprf("You can't mount anything outside of your normal form.");
-        return spret::abort;
-    }
-    else if (you.form != transformation::none)
-    {
-        if (!crawl_state.disables[DIS_CONFIRMATIONS] && 
-            !yesno("You have to untransform in order to properly ride a mount. Continue?", true, 0))
-        {
-            canned_msg(MSG_OK);
-            return spret::abort;
-        }
-        untransform();
-    }
-
-    fail_check();
-    int dur = 0;
-    int mount_hp = 0;
-    bool were_flying = you.airborne();
-    bool already_mount = you.mounted();
-
-    switch (mount)
-    {
-    case mount_type::drake:
-        mount_hp = 28 + random2(10);
-        break;
-    case mount_type::hydra:
-        mount_hp = 60 + random2(12);
-        break;
-    default:
-    case mount_type::spider:
-        mount_hp = 47 + random2(9);
-        break;
-    }
-
-    if (pow == 0)
-    {
-        dur = 100 + you.skill(SK_INVOCATIONS) * 2 + random2(you.skill(SK_INVOCATIONS));
-        mount_hp *= 10 + you.skill(SK_INVOCATIONS);
-        mount_hp = div_rand_round(mount_hp, 10);
-    }
-    else
-    {
-        dur = 100 + pow/2 + random2(pow/2);
-        mount_hp *= 100 + pow;
-        mount_hp = div_rand_round(mount_hp, 100);
-    }
-
-    if (you.mount != mount)
-    {
-        if (mount == mount_type::hydra)
-        {
-            you.mount_heads = 4 + max(0, div_round_up(pow - 50, 13));
-        }
-        else
-            you.mount_heads = 1;
-    }
-
-    if (already_mount)
-    {
-        if (you.mount == mount)
-        {
-            bool extend = you.duration[DUR_MOUNTED] < dur;
-            bool heal = (you.mount_hp != you.mount_hp_max) || (mount_statuses() > 0);
-            if (!extend)
-            {
-                if (!heal)
-                    mprf(MSGCH_DURATION, "You fail to extend your %s's duration any further.", you.mount_name().c_str());
-                else
-                    mprf(MSGCH_DURATION, "You heal your %s's injuries.", you.mount_name().c_str());
-            }
-            else
-            {
-                if (!heal)
-                    mprf(MSGCH_DURATION, "You increase the time your %s has in this world.", you.mount_name().c_str());
-                else
-                    mprf(MSGCH_DURATION, "You heal your %s and increase the time it has in this world.", you.mount_name().c_str());
-            }
-            cure_mount_debuffs();
-            you.mount_hp = you.mount_hp_max;
-        }
-        else
-        {
-            mprf(MSGCH_DURATION, "You dismiss your previous mount to call another one.");
-            dismount();
-            you.mount = mount;
-            you.mount_hp = you.mount_hp_max = mount_hp;
-        }
-    }
-    else
-    {
-        you.mount = mount;
-        you.mount_hp = you.mount_hp_max = mount_hp;
-        you.mount_energy = 10;
-    }
-
-    dur = max(you.duration[DUR_MOUNTED], dur);
-    you.set_duration(DUR_MOUNTED, dur, 200);
-
-    if (!already_mount)
-    {
-        mprf(MSGCH_DURATION, "You summon a %s and ride upon its back.", you.mount_name().c_str());
-
-        if (were_flying)
-            land_player();
-
-        redraw_screen();
-    }
-
-    return spret::success;
-}
