+++ crawl-ref/source/spl-other.h
@@ -44,4 +44,3 @@ spret cast_passwall(const coord_def& delta, int pow, bool fail);
 spret cast_intoxicate(int pow, bool fail);
 spret cast_darkness(int pow, bool fail);
 spret cast_SMD(const coord_def& target, int pow, bool fail);
-spret gain_mount(mount_type mount, int pow, bool fail);
